+++ tests/api/test_check.py
@@ -25,10 +25,32 @@ def test_unknown_field(rc: RawConfig):
     app = create_test_client(rc, raise_server_exceptions=False)
     app.authmodel('', ['check'])
     csv_manifest = convert_ascii_manifest_to_csv('''
-    id | d | r | b | m | property | typo
-       | datasets/gov/example     |
-       |   |   |   | Country      |
-       |   |   |   |   | name     | string
+    id | d | r | b | m | property | typo    | ref
+       | datasets/gov/example     |         |
+       |   |   |   | Country      |         |
+       |   |   |   |   | name     | string  |
+    ''')
+    resp = app.post('/:check', files={
+        'manifest': ('manifest.csv', csv_manifest, 'text/csv'),
+    })
+    assert resp.json() == {
+        'errors': [
+            {
+                'code': 'TabularManifestError',
+                'message': 'manifest.csv:1: Unknown columns: typo.',
+            },
+        ]
+    }
+
+
+def test_strip_unknown_field(rc: RawConfig):
+    app = create_test_client(rc, raise_server_exceptions=False)
+    app.authmodel('', ['check'])
+    csv_manifest = convert_ascii_manifest_to_csv('''
+    id | d | r | b | m | property | typo | type
+       | datasets/gov/example     |      |
+       |   |   |   | Country      |      |
+       |   |   |   |   | name     |      | string
     ''')
     resp = app.post('/:check', files={
         'manifest': ('manifest.csv', csv_manifest, 'text/csv'),